Remove EventLoop version of paginate functions (#550)
* Remove EventLoop paginate

* Remove Async from PaginateTests

* Remove @available(macOS...
